Complete dev part. Ready to test building
