+++ Testing/Temporary/LastTest.log
@@ -1,0 +1,3 @@
+Start testing: Dec 27 11:08 +03
+----------------------------------------------------------
+End testing: Dec 27 11:08 +03
