+++ compile.sh
@@ -3,5 +3,5 @@
 mkdir build && cd "$_"
 cmake -DBUILD_TESTING=ON ..
 cmake --build . -- -j2
-#ctest -j2
-#cpack .
+ctest -j2
+cpack .
