+++ custom_cmake/functions.cmake
@@ -1,12 +1,4 @@
 function(custom_compiler_opts TARGET)
-    target_compile_features(${TARGET}
-	    PUBLIC
-			cxx_std_14
-			cxx_variadic_templates
-			cxx_nullptr
-		PRIVATE
-			cxx_lambdas
-	)
     set_target_properties(${TARGET} PROPERTIES
 	    CXX_STANDARD 14
 	    CXX_STANDARD_REQUIRED ON
