--- misc/shared_ptr_alloc.h
@@ -1,14 +1,0 @@
-//
-// Created by dmitry on 12/25/17.
-//
-
-#ifndef ALLOCATOR_SHARED_PTR_ALLOC_H
-#define ALLOCATOR_SHARED_PTR_ALLOC_H
-
-
-class shared_ptr_alloc {
-
-};
-
-
-#endif //ALLOCATOR_SHARED_PTR_ALLOC_H
